Fix typo in Item-Spare.db2 structure

+++ WowPacketParser/Store/Objects/ItemTemplate.cs
@@ -38,7 +38,7 @@ namespace WowPacketParser.Store.Objects
         public float Unk430_2;
 
         [DBFieldName("Unk530_1", ClientVersionBuild.V5_3_0_16981)]
-        public float Unk530_1;
+        public int Unk530_1;
 
         [DBFieldName("BuyCount")]
         public uint BuyCount;
